FileUtils refactoring, so that it can be injected more flexibly if and when required.
