+++ generator/src/main/java/com/scottlogic/deg/generator/utils/FileUtils.java
@@ -4,134 +4,20 @@ import com.scottlogic.deg.generator.generation.GenerationConfigSource;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.FileAlreadyExistsException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.text.DecimalFormat;
-import java.util.Arrays;
-import java.util.function.Function;
 
-public class FileUtils {
-    /**
-     * Provides a decimal formatter for pre-padding numbers with just enough zeroes.
-     * E.g. if there are 42 records in a dataset this will produce a formatter of "00"
-     *
-     * @param numberOfDataSets Maximum number of data sets we will number
-     * @return Decimal format comprised of n zeroes where n is the number of digits in the input integer.
-     */
-    public static DecimalFormat getDecimalFormat(int numberOfDataSets) {
-        int maxNumberOfDigits = (int) Math.ceil(Math.log10(numberOfDataSets));
+public interface FileUtils {
+    boolean containsInvalidChars(File file);
 
-        char[] zeroes = new char[maxNumberOfDigits];
-        Arrays.fill(zeroes, '0');
+    boolean isFileEmpty(File file);
 
-        return new DecimalFormat(new String(zeroes));
-    }
+    boolean exists(Path path);
 
-    public boolean containsInvalidChars(File file) {
-        return file.toString().matches(".*[?%*|><\"].*|^(?:[^:]*+:){2,}[^:]*+$");
-    }
+    boolean isDirectory(Path path);
 
-    public boolean isFileEmpty(File file) {
-        return file.length() == 0;
-    }
+    boolean isDirectoryEmpty(Path filepath, int fileCount);
 
-    /**
-     * @param path the Path to check the existence of.
-     * @return true if the supplied FileOutputTarget exists on disk, false otherwise.
-     */
-    public boolean exists(Path path) {
-        return Files.exists(path);
-    }
+    boolean createDirectories(Path dir) throws IOException;
 
-    /**
-     * @param path the Path to test.
-     * @return true is the supplied FileOutputTarget is a directory, false otherwise.
-     */
-    public boolean isDirectory(Path path) {
-        return Files.isDirectory(path);
-    }
-
-    /**
-     * check for the existence of files named "<code>manifest.json</code>"
-     * and "<code>/^[0-9]{filecount}.csv$/</code>" in the given directory.
-     *
-     * @param filepath    the Path that contains the directory to test.
-     * @param fileCount the number of files we will check for.
-     * @return true if any of the files exist, false only if none of the files exist in the directory.
-     */
-    public boolean isDirectoryEmpty(Path filepath, int fileCount) {
-        if (directoryContainsManifestJsonFile(filepath) ||
-            directoryContainsFilesWithExt(filepath, "csv", fileCount)) {
-            return false;
-        }
-        return true;
-    }
-
-    private boolean directoryContainsManifestJsonFile(Path filePath) {
-        return Files.exists(Paths.get(filePath.toString(), "manifest.json"));
-    }
-
-    private boolean directoryContainsFilesWithExt(Path filePath, String ext, int fileCount) {
-        DecimalFormat intFormatter = FileUtils.getDecimalFormat(fileCount);
-        for (int x = 1; x <= fileCount; x++) {
-            if (Files.exists(Paths.get(filePath.toString(), intFormatter.format(x) + "." + ext))) {
-                return true;
-            }
-        }
-        return false;
-    }
-
-    public File getTraceFile(GenerationConfigSource configSource) {
-        String filenameWithoutExtension = configSource.getOutputPath().toString().replaceAll("\\.[^.]+$", "");
-        return Paths.get(filenameWithoutExtension + "-trace.json").toFile();
-    }
-
-    /**
-     * we wrap Files.createDirectories() so that we can mock it out in unit tests.
-     *
-     * @param dir the directory we want to create
-     * @return true is the directory was created successfully,
-     * false if the parent already exists and is a file
-     * @throws IOException if we are unable to create the directory due to an I/O error
-     */
-    public boolean createDirectories(Path dir) throws IOException {
-        try {
-            checkValidDirectoryPath(dir);
-            Files.createDirectories(dir);
-        } catch (FileAlreadyExistsException e) {
-            return false;
-        }
-        return true;
-    }
-
-    private void checkValidDirectoryPath(Path dir) throws FileAlreadyExistsException {
-        if (Files.exists(dir) && !Files.isDirectory(dir)) {
-            throw new FileAlreadyExistsException("Directory name exists as file");
-        }
-        Path prntDir = dir.getParent();
-        if (prntDir != null) {
-            checkValidDirectoryPath(prntDir);
-        }
-    }
-
-    /** addFilenameSuffix("C:\something.bat", "-trace") == "C:\something-trace.bat" */
-    public static Path addFilenameSuffix(Path path, String newSuffix) {
-        return transformFilename(
-            path,
-            filename -> filename.replaceAll("\\.[^.]+$", newSuffix + "$0"));
-    }
-
-    public static Path replaceExtension(Path path, String newExtension) {
-        return transformFilename(
-            path,
-            filename -> filename.replaceAll("\\.[^.]+$", "." + newExtension));
-    }
-
-    private static Path transformFilename(Path filePath, Function<String, String> transformFunc) {
-        final String newFilename = transformFunc.apply(filePath.getFileName().toString());
-
-        return filePath.resolveSibling(Paths.get(newFilename));
-    }
+    File getTraceFile(GenerationConfigSource configSource);
 }
